Basic Page Fault Support! + Comments + Fixes
Added some comments on things that need to get done, and removed some comments on things that are already done.

Fixed another small bug that caused certain single step exceptions to not be caught and handled.

This is a working version of the program with basic page fault support for hooks.